A new Logic was written for using default sku inventory instead of alt sku inventory
Link to QA issue
https://github.com/BroadleafCommerce/QA/issues/4674
