+++ admin/broadleaf-open-admin-platform/src/main/resources/messages/GeneratedMessagesEntityFramework.properties
@@ -75,8 +75,8 @@ ProductImpl_MetaDescription=Metadata Description Override
 ProductImpl_MetaDescription_Tooltip=Use this field to set the page description while viewing this product.
 ProductImpl_Canonical_Url=Canonical URL
 ProductImpl_Canonical_Url_Tooltip=Use this field to set the canonical URL for the page.
-ProductImpl_useDefaultSkuInInventory=Default sku in inventory.
-ProductImpl_useDefaultSkuInInventory_Tooltip=Use default sku in inventory.
+ProductImpl_useDefaultSkuInInventory=Track Inventory by Product.
+ProductImpl_useDefaultSkuInInventory_Tooltip=Uses the Product to track inventory instead of the Sku.
 ProductWeight_Product_Weight=Weight
 ProductWeight_Product_Weight_Units=Weight Units
 SkuImpl_Sku_ID=Sku ID
