+++ core/broadleaf-framework/src/main/java/org/broadleafcommerce/core/catalog/domain/Product.java
@@ -881,10 +881,6 @@ public interface Product extends Serializable, MultiTenantCloneable<Product>, In
      */
     public boolean isOnSale();
 
-    public Boolean getUseDefaultSkuInInventory();
-
-    public void setUseDefaultSkuInInventory(Boolean useDefaultSkuInInventory);
-
     /**
      * added just for convenience, references defaultSku.margin
      * @return
