+++ core/broadleaf-framework/src/main/java/org/broadleafcommerce/core/inventory/service/InventoryServiceImpl.java
@@ -28,6 +28,7 @@ import org.broadleafcommerce.common.event.BroadleafSystemEventDetail;
 import org.broadleafcommerce.common.extension.ExtensionResultHolder;
 import org.broadleafcommerce.common.extension.ExtensionResultStatusType;
 import org.broadleafcommerce.common.util.TransactionUtils;
+import org.broadleafcommerce.core.catalog.domain.ProductSkuUsage;
 import org.broadleafcommerce.core.catalog.domain.Sku;
 import org.broadleafcommerce.core.catalog.service.CatalogService;
 import org.broadleafcommerce.core.inventory.service.type.InventoryType;
@@ -37,6 +38,7 @@ import org.broadleafcommerce.core.order.domain.Order;
 import org.broadleafcommerce.core.order.domain.OrderItem;
 import org.hibernate.annotations.Cache;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -65,6 +67,10 @@ public class InventoryServiceImpl implements ContextualInventoryService {
 
     @Autowired
     protected ApplicationContext applicationContext;
+
+    @Value("${enable.weave.use.default.sku.inventory:false}")
+    protected boolean enableUseDegaultSkuInventory = false;
+
 
     @Override
     public boolean checkBasicAvailablility(Sku sku) {
@@ -138,7 +144,7 @@ public class InventoryServiceImpl implements ContextualInventoryService {
 
             for (Sku sku : skus) {
                 Sku skuForInventory = sku;
-                if (sku.getProduct().getUseDefaultSkuInInventory()){
+                if (enableUseDegaultSkuInventory && ((ProductSkuUsage) sku.getProduct()).getUseDefaultSkuInInventory()){
                     skuForInventory = sku.getProduct().getDefaultSku();
                 }
                 Integer quantityAvailable = 0;
@@ -165,7 +171,7 @@ public class InventoryServiceImpl implements ContextualInventoryService {
     @Override
     public boolean isAvailable(Sku sku, int quantity, Map<String, Object> context) {
         Sku skuForInventory = sku;
-        if (sku.getProduct().getUseDefaultSkuInInventory()){
+        if (enableUseDegaultSkuInventory && ((ProductSkuUsage) sku.getProduct()).getUseDefaultSkuInInventory()){
             skuForInventory = sku.getProduct().getDefaultSku();
         }
         if (quantity < 1) {
@@ -205,7 +211,7 @@ public class InventoryServiceImpl implements ContextualInventoryService {
         for (Entry<Sku, Integer> entry : skuQuantities.entrySet()) {
             Sku sku = entry.getKey();
             Sku skuForInventory = sku;
-            if (sku.getProduct().getUseDefaultSkuInInventory()) {
+            if (enableUseDegaultSkuInventory && ((ProductSkuUsage) sku.getProduct()).getUseDefaultSkuInInventory()){
                 skuForInventory = sku.getProduct().getDefaultSku();
             }
             Integer quantity = entry.getValue();
@@ -225,9 +231,7 @@ public class InventoryServiceImpl implements ContextualInventoryService {
                     }
                     int newInventory = inventoryAvailable - quantity;
                     skuForInventory.setQuantityAvailable(newInventory);
-                    catalogService.saveSku(sku);
                     catalogService.saveSku(skuForInventory);
-                    invalidateSkuInventory(sku);
                     invalidateSkuInventory(skuForInventory);
                 } else {
                     LOG.info("Not decrementing inventory as the Sku has been marked as always available");
@@ -258,8 +262,9 @@ public class InventoryServiceImpl implements ContextualInventoryService {
     protected void incrementSku(Map<Sku, Integer> skuQuantities, Map<String, Object> context) {
         for (Entry<Sku, Integer> entry : skuQuantities.entrySet()) {
             Sku sku = entry.getKey();
+
             Sku skuForInventory = sku;
-            if (sku.getProduct().getUseDefaultSkuInInventory()) {
+            if (enableUseDegaultSkuInventory && ((ProductSkuUsage) sku.getProduct()).getUseDefaultSkuInInventory()){
                 skuForInventory = sku.getProduct().getDefaultSku();
             }
             Integer quantity = entry.getValue();
@@ -273,9 +278,7 @@ public class InventoryServiceImpl implements ContextualInventoryService {
                 }
                 int newInventory = currentInventoryAvailable + quantity;
                 skuForInventory.setQuantityAvailable(newInventory);
-                catalogService.saveSku(sku);
                 catalogService.saveSku(skuForInventory);
-                invalidateSkuInventory(sku);
                 invalidateSkuInventory(skuForInventory);
             } else {
                 LOG.info("Not incrementing inventory as the Sku has been marked as always available");
@@ -306,7 +309,7 @@ public class InventoryServiceImpl implements ContextualInventoryService {
             if (orderItem instanceof DiscreteOrderItem) {
                 Sku sku = ((DiscreteOrderItem) orderItem).getSku();
                 Sku skuForInventory = sku;
-                if (sku.getProduct().getUseDefaultSkuInInventory()) {
+                if (enableUseDegaultSkuInventory && ((ProductSkuUsage) sku.getProduct()).getUseDefaultSkuInInventory()){
                     skuForInventory = sku.getProduct().getDefaultSku();
                 }
                 Integer quantity = skuInventoryMap.get(skuForInventory);
@@ -321,7 +324,7 @@ public class InventoryServiceImpl implements ContextualInventoryService {
             } else if (orderItem instanceof BundleOrderItem) {
                 BundleOrderItem bundleItem = (BundleOrderItem) orderItem;
                 Sku bundleSku = bundleItem.getSku();
-                if (bundleSku.getProduct().getUseDefaultSkuInInventory()) {
+                if (enableUseDegaultSkuInventory && ((ProductSkuUsage) bundleSku.getProduct()).getUseDefaultSkuInInventory()){
                     bundleSku = bundleSku.getProduct().getDefaultSku();
                 }
                 if (InventoryType.CHECK_QUANTITY.equals(bundleSku.getInventoryType())) {
@@ -333,7 +336,7 @@ public class InventoryServiceImpl implements ContextualInventoryService {
                 List<DiscreteOrderItem> discreteItems = bundleItem.getDiscreteOrderItems();
                 for (DiscreteOrderItem discreteItem : discreteItems) {
                     Sku sku = discreteItem.getSku();
-                    if (sku.getProduct().getUseDefaultSkuInInventory()) {
+                    if (enableUseDegaultSkuInventory && ((ProductSkuUsage) sku.getProduct()).getUseDefaultSkuInInventory()){
                         sku = sku.getProduct().getDefaultSku();
                     }
                     if (InventoryType.CHECK_QUANTITY.equals(sku.getInventoryType())) {
@@ -412,7 +415,7 @@ public class InventoryServiceImpl implements ContextualInventoryService {
     @Override
     public void checkSkuAvailability(Order order, Sku sku, Integer requestedQuantity) throws InventoryUnavailableException {
         Sku skuForInventory = sku;
-        if (sku.getProduct().getUseDefaultSkuInInventory()) {
+        if (enableUseDegaultSkuInventory && ((ProductSkuUsage) sku.getProduct()).getUseDefaultSkuInInventory()){
             skuForInventory = sku.getProduct().getDefaultSku();
         }
         // First check if this Sku is available
